Create parent component to Game component
Will be used to conditionally render either the Game or a message for people on mobile devices / small screens
